pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "thumbnail-generator"
version = "1.0.0"
description = "Generate thumbnails for images and videos following freedesktop.org standards"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["thumbnail", "image", "video", "nas", "freedesktop"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "Pillow>=10.0.0",
]

[project.scripts]
thumbnail-generator = "thumbnail_generator:main"

[project.urls]
Homepage = "https://github.com/yourusername/thumbnail_generator"
Repository = "https://github.com/yourusername/thumbnail_generator"
Issues = "https://github.com/yourusername/thumbnail_generator/issues"
